hack/make-rules: reimplement go2make for modules
Strips the flags and help, because this is a first proof of concept.

The basic idea is to list all packages and their transitive deps in the
main module, which uses the vendor folder. Then, for each k8s.io package
found, list all the files for itself and its direct dependencies.

Two open questions remain, which are left as TODOs.

Overall, this program runs in ~3s for me and outputs 11MiB of text,
while the old program in GOPATH mode ran in ~4s and output 13MiB. The
output is not an exact match since the old program also produced
definitions for non-k8s.io packages, which presumably was unnecessary.

The entire pkgwalk code is now unnecessary, since 'go list -deps' does
that work for us. golist.go contains a thin wrapper for its JSON
interface; see 'go help list'.